moved generate helpers to utils
